+++ apps/freelance/src/pages/Freelancer/SearchWork/style.tsx
@@ -1,32 +1,5 @@
 import styled from "styled-components";
 
-const StyledPage = styled.div`
-	margin: auto;
-	display: flex;
-	height: 100vh;
-
-	background: ${({ theme }) => theme.signPages.image}, ${({ theme }) => theme.colors.lightGrey};
-
-	* {
-		scrollbar-width: auto;
-		scrollbar-color: ${({ theme }) => theme.colors.lightRed} ${({ theme }) => theme.colors.white};
-	}
-
-	*::-webkit-scrollbar {
-		width: 15px;
-	}
-
-	*::-webkit-scrollbar-track {
-		background: ${({ theme }) => theme.colors.white};
-	}
-
-	*::-webkit-scrollbar-thumb {
-		background-color: ${({ theme }) => theme.colors.lightRed};
-		border-radius: 10px;
-		border: 3px solid ${({ theme }) => theme.colors.white};
-	}
-`;
-
 export const InputWrapper = styled.div`
 	display: flex;
 	flex-direction: column;
@@ -34,5 +7,3 @@ export const InputWrapper = styled.div`
 	width: 100%;
 	height: 100%;
 `;
-
-export { StyledPage };
