+++ libs/components/src/lib/filter-container/filter-container.styled.tsx
@@ -5,16 +5,15 @@ interface IFilterContainer {
 }
 
 export const FilterContainer = styled.div<IFilterContainer>`
-    padding: 2rem;
+    padding: 1.5rem;
     border-radius: 15px;
-    border:  1px solid ${({ theme }) => theme.colors.lightRed};
     position: absolute;
-    top: 16%;
-    right: 6%;
-    width: 30%;
-    background-color: ${({ theme }) => theme.colors.lightGrey};
+    top: 17%;
+    right: 5.9%;
+    width: 27%;
+    background-color: rgb(229, 218, 218, 0.85);
     transition: 0.5s;
     transform: translateX(120%);
     
-    ${({ isShown }) => isShown && css`transform: translateX(20%)`}
+    ${({ isShown }) => isShown && css`transform: translateX(13%)`}
 `;
